requirements.txt
# Core dependencies
python-dotenv==1.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
pandas>=2.1.0

# Dataset
kagglehub[pandas-datasets]

# Vector database
qdrant-client>=1.7.0

# Embedding models
sentence-transformers>=2.5.0
transformers>=4.40.0
torch>=2.2.0

# LLM integration
langchain>=0.1.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0

# Utils
tqdm==4.66.1
